Renamed push/pull commands, worked on push
